embeddings/src/tools-service/requirements.txt
fastapi
uvicorn
transformers
accelerate
sentencepiece
protobuf
pillow
librosa